requirements.txt
black==22.12.0
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
idna==3.4
mypy-extensions==0.4.3
pathspec==0.11.0
platformdirs==2.6.2
prometheus-client==0.16.0
PyYAML==6.0
requests==2.28.2
slack-sdk==3.19.5
tomli==2.0.1
urllib3==1.26.14
